TESTING_SEEK/unilabs-referencias-bend/requirements_fixed.txt
amqp==5.1.1
arabic-reshaper==2.1.4
asgiref==3.3.4
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.1
django-admin-autocomplete-filter==0.7.1
django-admin-inline-paginator==0.2.2
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.2
django-admin-sortable==2.3
django-appconf==1.0.5
django-cors-headers==3.10.0
django-csp==3.7
django-permissions-policy==4.9.0
djangorestframework==3.12.4
djangorestframework-recaptcha==0.2.0
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.2.1
drf-renderer-xlsx==0.4.4
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
importlib-metadata==4.11.4
kombu==5.2.4
MarkupPy==1.14
numpy==1.21.5
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==3.0.9
PyPDF2==1.26.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2022.1
PyYAML==5.4.1
redis==4.3.3
reportlab==3.6.3
requests==2.25.1
requests-toolbelt==0.9.1
sentry-sdk==1.5.12
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
typing_extensions==4.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
django-simple-history==3.0.0